requirements.txt
cryptography==2.8
Django==2.2.10
mysqlclient==1.4.6
djangorestframework==3.11.0
django-rest-swagger==2.2.0
django-filter==2.2.0
drf-nested-routers==0.91
gunicorn==20.0.4
django-multiselectfield==0.1.12
django-mysql==3.3.0
django-jet==1.0.8
django-ckeditor==5.9.0
drf_yasg==1.17.1
django-crispy-forms==1.8.1
mysql-connector-python==8.0.18
django-fernet-fields==0.6
jira==2.0.0
django-sitetree==1.14.0
django-jsonfield==1.4.0
